scraper/requirements.txt
discord==2.1.0
python-dotenv==0.21.1
urlextract==1.8.0
libsql-client==0.1.2
modal-client==0.45.1163
